// src/app/models/patient.model.ts
export interface Patient {
    // id: string;
    name: string;
    vaccineName: string;
    dueDate: string;
    givenOn: string;
    brandGiven: string;
    dose: string;
    dob: string;
    gender: string;
    birthLocation: string;
    bloodGroup: string;
    weight: string;
    height: string;
}